services/negotiation/listeners/gemini_realtime.py
"""Gemini realtime listener adapter for bidirectional audio streaming."""

import asyncio
import json
import websockets
from typing import AsyncIterator, Dict, Any
from .base import Listener


class GeminiRealtimeListener(Listener):
    """Gemini realtime listener using WebSocket streaming.

    Note: This is a stub implementation. Gemini Live API uses gRPC.
    For a full implementation, you'd need:
    1. Gemini Live API credentials
    2. gRPC client setup
    3. Proper audio encoding (mu-law)
    """

    def __init__(self, config: Dict[str, Any]):
        super().__init__(config)
        self.api_key = config.get("api_key")
        self.model = config.get("model", "models/gemini-1.5-flash")
        self.websocket = None
        self.event_queue = asyncio.Queue()
        self.running = False

    async def start(self) -> None:
        """Start the Gemini realtime connection."""
        if not self.api_key:
            print("GEMINI_API_KEY not set. Using mock mode.")
            self.running = True
            asyncio.create_task(self._mock_stream())
            return

        try:
            # Note: Gemini Live API uses gRPC, not WebSocket
            # This is a placeholder for the actual implementation
            self.running = True
            asyncio.create_task(self._mock_stream())
            print("Gemini realtime listener started (stub mode)")
        except Exception as e:
            print(f"Error starting Gemini realtime: {e}")

    async def stop(self) -> None:
        """Stop the connection."""
        self.running = False
        if self.websocket:
            await self.websocket.close()
        if self.event_queue:
            await self.event_queue.put({"type": "stop"})

    async def feed_pcm(self, pcm_bytes: bytes, ts_ms: int) -> None:
        """Feed PCM audio to Gemini."""
        if not self.running:
            return

        # Note: In real implementation, convert PCM to mu-law and send via gRPC
        # For now, we simulate processing
        pass

    async def final_text(self) -> str:
        """Get final recognized text."""
        return "Final transcript from Gemini realtime"  # Placeholder

    async def stream_events(self) -> AsyncIterator[Dict[str, Any]]:
        """Stream events from Gemini."""
        while self.running:
            try:
                event = await asyncio.wait_for(self.event_queue.get(), timeout=0.1)
                if event.get("type") == "stop":
                    break
                yield event
            except asyncio.TimeoutError:
                continue
            except Exception as e:
                print(f"Error in stream_events: {e}")
                break

    async def _mock_stream(self) -> None:
        """Mock streaming for testing."""
        while self.running:
            try:
                await asyncio.sleep(1.0)
                # Simulate subtitle events
                await self.event_queue.put({
                    "type": "subtitle",
                    "text": "Processing audio with Gemini realtime...",
                    "final": False,
                    "confidence": 0.9
                })
            except Exception as e:
                print(f"Error in mock stream: {e}")


# TODO: Real Gemini Live implementation
# The actual Gemini Live API requires:
# 1. OAuth2 credentials and project setup
# 2. gRPC client with proper authentication
# 3. Audio encoding to mu-law format
# 4. Bidirectional streaming with proper session management

# Example of what the real implementation would look like:
"""
import grpc
from google.auth.transport.grpc import AuthMetadataPlugin
from google.oauth2 import service_account

class RealGeminiRealtimeListener(Listener):
    async def start(self):
        # Set up OAuth2 credentials
        credentials = service_account.Credentials.from_service_account_file(
            self.config.get("credentials_file", "service_account.json")
        )

        # Create gRPC channel with auth
        auth_plugin = AuthMetadataPlugin(credentials)
        channel = grpc.secure_channel(
            "speech.googleapis.com:443",
            grpc.composite_channel_credentials(
                grpc.ssl_channel_credentials(),
                grpc.auth._google_auth_interceptor.AuthMetadataInterceptor(auth_plugin)
            )
        )

        # Create stub and start session
        stub = speech_pb2_grpc.SpeechStub(channel)
        # ... session setup code ...

    async def feed_pcm(self, pcm_bytes: bytes, ts_ms: int):
        # Convert to mu-law and stream to Gemini
        # ... audio processing ...
        pass
"""


# Re-export for convenience
__all__ = ["GeminiRealtimeListener"]
